feat: Add Alt+Click drag to adjust viewport
This commit introduces the ability for users to adjust the camera's rotation and tilt by holding the Alt key and dragging the mouse on the map canvas.

Key changes:
- Implemented mouse event handlers in `MapCanvas` to track Alt+Drag gestures.
- Horizontal mouse movement now controls the camera's yaw (rotation).
- Vertical mouse movement now controls the camera's pitch (tilt).
- The manual tilt functionality is disabled when the "Auto-tilt" feature is enabled to prevent conflicting behaviors.
- The allowed range for horizontal rotation (`horizontalAngle`) has been expanded from +/-45 to +/-180 degrees in the configuration to allow for a full 360-degree rotation.
- A new `AltDragState` struct has been added to `MapCanvas` to cleanly manage the state of the drag operation.
- The adjusted camera angles are automatically saved to the application settings, persisting the viewport state between sessions.
- The mouse cursor now changes to a "closed hand" icon during the Alt+Drag operation to provide clear visual feedback to the user.
- Fixed a compile-time error related to implicit float conversions that was caught by the CI build.

+++ src/display/mapcanvas.cpp
@@ -607,7 +607,8 @@ void MapCanvas::mouseMoveEvent(QMouseEvent *const event)
         // Horizontal movement adjusts yaw (horizontalAngle)
         const int dx = delta.x();
         if (dx != 0) {
-            conf.horizontalAngle.set(conf.horizontalAngle.get() + static_cast<int>(dx * SENSITIVITY));
+            conf.horizontalAngle.set(conf.horizontalAngle.get()
+                                       + static_cast<int>(static_cast<float>(dx) * SENSITIVITY));
         }
 
         // Vertical movement adjusts pitch (verticalAngle), if auto-tilt is off
@@ -615,8 +616,9 @@ void MapCanvas::mouseMoveEvent(QMouseEvent *const event)
             const int dy = delta.y();
             if (dy != 0) {
                 // Negated to match intuitive up/down dragging
-                conf.verticalAngle.set(conf.verticalAngle.get()
-                                       + static_cast<int>(-dy * SENSITIVITY));
+                conf.verticalAngle.set(
+                    conf.verticalAngle.get()
+                    + static_cast<int>(static_cast<float>(-dy) * SENSITIVITY));
             }
         }
